Move CQRS request interfaces to a more appropriate file
