+++ BattleshipGame.WebApi/Contracts/v1/Requests/CancelGame/CancelGameCommandRequest.cs
@@ -1,4 +1,4 @@
-using BattleshipGame.WebApi.RequestProcessor;
+using frm.Infrastructure.Cqrs.Requests;
 using Microsoft.AspNetCore.Mvc;
 
 namespace BattleshipGame.WebApi.Contracts.v1.Requests.CancelGame;
