+++ BattleshipGame.WebApi/Controllers/v1/GameController.cs
@@ -1,8 +1,10 @@
 using System.Net.Mime;
+using Asp.Versioning;
 using BattleshipGame.WebApi.Contracts.v1.Requests.CancelGame;
 using BattleshipGame.WebApi.Contracts.v1.Requests.Fire;
 using BattleshipGame.WebApi.Contracts.v1.Requests.InitGame;
-using BattleshipGame.WebApi.RequestProcessor;
+using BattleshipGame.WebApi.Contracts.v1.Responses;
+using frm.Infrastructure.Cqrs.Requests;
 using Microsoft.AspNetCore.Mvc;
 
 namespace BattleshipGame.WebApi.Controllers.v1;
