+++ frm.Infrastructure.Cqrs/Requests/ICommandRequestHandler.cs
@@ -1,4 +1,4 @@
-namespace BattleshipGame.WebApi.RequestProcessor;
+namespace frm.Infrastructure.Cqrs.Requests;
 
 public interface ICommandRequestHandler<in TRequest, TResponse> where TRequest : ICommandRequest<TResponse>
 {
